requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
pycparser==3.0
PyJWT==2.12.1
sqlparse==0.5.5
